+++ tests/ui/linking/crate-type-invalid-flag-error.stderr
